added more logs in animation systems

+++ src/engine/resources/all/Animation.cpp
@@ -8,7 +8,7 @@
 #include "Animation.hpp"
 #include "Logging.hpp"
 
-Animation::Animation(const std::vector<std::shared_ptr<Texture>>& frames, double duration, bool loop, const std::string& name):
+Animation::Animation(const std::vector<std::shared_ptr<Texture>>& frames, float duration, bool loop, const std::string& name):
     _frames(frames),
     _duration(duration),
     _loop(loop),
